Server/requirements.txt
PyMySQL==0.9.2
flask==1.0.2
Jinja2==2.10
flask_restful==0.3.6
python-swiftclient==3.6.0
python-keystoneclient==3.18.0